finished cart checkout

+++ controllers/ingredients.js
@@ -44,7 +44,7 @@ exports.addIngredients = (req,res) => {
             } else{
                 // console.log(resp);
                 var oldAmount = resp.rows[0].amount;
-                var newAmount = parseInt(oldAmount,10) + parseInt(amt,10);
+                var newAmount = parseFloat(oldAmount,10) + parseFloat(amt,10);
                 var update_user_ing =  'UPDATE has SET amount = $1 WHERE username=$2 and ingredient_id = $3';
 
                 pool.query(update_user_ing,[newAmount,username,ing_id],(error,results) => {
